feat: updated styling and allowed for navigation through help menu and is ready for info
