+++ apps/frontend/src/routes/PatientGuidePage.tsx
@@ -1,7 +1,21 @@
+import {
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+} from "@/components/ui/accordion.tsx";
+
 export default function PatientGuidePage() {
   return (
     <>
-      <p>Patient</p>
+      <Accordion type="single" collapsible className="w-auto col-span-3 mx-10">
+        <AccordionItem value="Intro">
+          <AccordionTrigger className="text-black">
+            Patient Features
+          </AccordionTrigger>
+          <AccordionContent>*Put Intro Here*</AccordionContent>
+        </AccordionItem>
+      </Accordion>
     </>
   );
 }
